+++ locals.tf
@@ -29,26 +29,23 @@ locals {
   s3_buckets                                = concat(local.s3_instance_buckets, [aws_s3_bucket.bucket_logs.bucket], local.license_server_bucket)
   private_subnets                           = local.create_vpc ? module.vpc[0].private_subnets : (local.use_private_subnets_ids ? var.private_subnet_ids : [for s in data.aws_subnet.private_subnet : s.id])
   public_subnets                            = local.create_vpc ? module.vpc[0].public_subnets : (local.use_public_subnet_ids ? var.public_subnet_ids : [for s in data.aws_subnet.public_subnet : s.id])
-  # Using a one-line command for gpuPostUserData to avoid issues due to different line endings between Windows and Linux.
 
-  default_managed_node_pools = {
+  default_node_pools = {
     "default" = {
       node_group_name = "default"
       instance_types  = var.linuxNodeSize
       subnet_ids      = local.private_subnets
-      desired_size    = var.linuxNodeCountMin
       max_size        = var.linuxNodeCountMax
       min_size        = var.linuxNodeCountMin
-      disk_size       = var.linuxNodeDiskSize
+      volume_size     = var.linuxNodeDiskSize
     },
     "execnodes" = {
       node_group_name = "execnodes"
       instance_types  = var.linuxExecutionNodeSize
       subnet_ids      = local.private_subnets
-      desired_size    = var.linuxExecutionNodeCountMin
       max_size        = var.linuxExecutionNodeCountMax
       min_size        = var.linuxExecutionNodeCountMin
-      disk_size       = var.linuxExecutionNodeDiskSize
+      volume_size     = var.linuxExecutionNodeDiskSize
       k8s_labels = {
         "purpose" = "execution"
       }
@@ -61,24 +58,16 @@ locals {
       ]
     }
   }
-
   gpu_node_pool = {
     "gpuexecnodes" = {
-      node_group_name        = "gpuexecnodes"
-      instance_types         = var.gpuNodeSize
-      subnet_ids             = local.private_subnets
-      desired_size           = var.gpuNodeCountMin
-      max_size               = var.gpuNodeCountMax
-      min_size               = var.gpuNodeCountMin
-      disk_size              = var.gpuNodeDiskSize
-      custom_ami_id          = data.aws_ami.al2gpu_ami.image_id
-      create_launch_template = true
-      block_device_mappings = [{
-        device_name           = "/dev/sda1"
-        volume_type           = "gp3"
-        volume_size           = 128
-        delete_on_termination = true
-      }]
+      node_group_name   = "gpuexecnodes"
+      instance_types    = var.gpuNodeSize
+      subnet_ids        = local.private_subnets
+      max_size          = var.gpuNodeCountMax
+      min_size          = var.gpuNodeCountMin
+      custom_ami_id     = data.aws_ami.al2gpu_ami.image_id
+      block_device_name = "/dev/sda1"
+      volume_size       = var.gpuNodeDiskSize
       k8s_labels = {
         "purpose" = "gpu"
       }
@@ -91,24 +80,16 @@ locals {
       ]
     }
   }
-
   ivsgpu_node_pool = {
     "gpuivsnodes" = {
-      node_group_name        = "gpuivsnodes"
-      instance_types         = var.ivsGpuNodeSize
-      subnet_ids             = local.private_subnets
-      desired_size           = var.ivsGpuNodeCountMin
-      max_size               = var.ivsGpuNodeCountMax
-      min_size               = var.ivsGpuNodeCountMin
-      disk_size              = var.ivsGpuNodeDiskSize
-      custom_ami_id          = data.aws_ami.al2gpu_ami.image_id
-      create_launch_template = true
-      block_device_mappings = [{
-        device_name           = "/dev/sda1"
-        volume_type           = "gp3"
-        volume_size           = 128
-        delete_on_termination = true
-      }]
+      node_group_name   = "gpuivsnodes"
+      instance_types    = var.ivsGpuNodeSize
+      subnet_ids        = local.private_subnets
+      max_size          = var.ivsGpuNodeCountMax
+      min_size          = var.ivsGpuNodeCountMin
+      custom_ami_id     = data.aws_ami.al2gpu_ami.image_id
+      block_device_name = "/dev/sda1"
+      volume_size       = var.ivsGpuNodeDiskSize
       k8s_labels = {
         "product" = "ivs",
         "purpose" = "gpu"
@@ -126,5 +107,14 @@ locals {
         }
       ]
     }
+  }
+  node_pools = merge(local.default_node_pools, var.gpuNodePool ? local.gpu_node_pool : {}, var.ivsGpuNodePool ? local.ivsgpu_node_pool : {})
+  aws_context = {
+    caller_identity_account_id = data.aws_caller_identity.current.account_id
+    partition_dns_suffix       = data.aws_partition.current.dns_suffix
+    partition_id               = data.aws_partition.current.id
+    partition                  = data.aws_partition.current.partition
+    region_name                = data.aws_region.current.name
+    iam_issuer_arn             = data.aws_iam_session_context.current.issuer_arn
   }
 }