+++ modules/eks/cluster-auth.tf
@@ -1,0 +1,28 @@
+resource "kubernetes_config_map" "aws_auth" {
+  metadata {
+    name      = "aws-auth"
+    namespace = "kube-system"
+    labels = merge(
+      {
+        "app.kubernetes.io/managed-by" = "terraform"
+        "app.kubernetes.io/created-by" = "terraform"
+      },
+    )
+  }
+
+  data = {
+    mapRoles = yamlencode(
+      distinct(concat(
+        local.node_group_aws_auth_config_map,
+        var.map_roles,
+      ))
+    )
+    mapUsers    = yamlencode(var.map_users)
+    mapAccounts = yamlencode(var.map_accounts)
+  }
+
+  depends_on = [
+    aws_eks_cluster.eks,
+    data.http.eks_cluster_readiness
+  ]
+}