+++ modules/eks/data.tf
@@ -1,0 +1,137 @@
+data "aws_eks_cluster" "cluster" {
+  name = aws_eks_cluster.eks.id
+}
+
+data "http" "eks_cluster_readiness" {
+  url                = join("/", [data.aws_eks_cluster.cluster.endpoint, "healthz"])
+  ca_cert_pem        = base64decode(data.aws_eks_cluster.cluster.certificate_authority[0].data)
+  request_timeout_ms = 600000
+}
+
+data "aws_iam_policy_document" "eks_key" {
+  statement {
+    sid    = "Allow access for all principals in the account that are authorized"
+    effect = "Allow"
+    actions = [
+      "kms:CreateGrant",
+      "kms:Decrypt",
+      "kms:DescribeKey",
+      "kms:Encrypt",
+      "kms:GenerateDataKey*",
+      "kms:ReEncrypt*",
+    ]
+    resources = ["*"]
+
+    principals {
+      type = "AWS"
+      identifiers = [
+        "arn:${var.aws_context.partition_id}:iam::${var.aws_context.caller_identity_account_id}:root"
+      ]
+    }
+
+    condition {
+      test     = "StringEquals"
+      variable = "kms:CallerAccount"
+      values   = [var.aws_context.caller_identity_account_id]
+    }
+
+    condition {
+      test     = "StringEquals"
+      variable = "kms:ViaService"
+      values   = ["eks.${var.aws_context.region_name}.amazonaws.com"]
+    }
+  }
+
+  statement {
+    sid    = "Allow direct access to key metadata to the account"
+    effect = "Allow"
+    actions = [
+      "kms:Describe*",
+      "kms:Get*",
+      "kms:List*",
+      "kms:RevokeGrant",
+    ]
+    resources = ["*"]
+
+    principals {
+      type = "AWS"
+      identifiers = [
+        "arn:${var.aws_context.partition_id}:iam::${var.aws_context.caller_identity_account_id}:root"
+      ]
+    }
+  }
+
+  statement {
+    sid    = "Allow access for Key Administrators"
+    effect = "Allow"
+    actions = [
+      "kms:*"
+    ]
+    resources = ["*"]
+
+    principals {
+      type        = "AWS"
+      identifiers = [var.aws_context.iam_issuer_arn]
+    }
+  }
+
+  statement {
+    sid    = "Allow use of the key"
+    effect = "Allow"
+    actions = [
+      "kms:Decrypt",
+      "kms:DescribeKey",
+      "kms:Encrypt",
+      "kms:GenerateDataKey*",
+      "kms:ReEncrypt*",
+    ]
+    resources = ["*"]
+
+    principals {
+      type = "AWS"
+      identifiers = [
+        local.cluster_iam_role_pathed_arn
+      ]
+    }
+  }
+
+  statement {
+    sid    = "Allow attachment of persistent resources"
+    effect = "Allow"
+    actions = [
+      "kms:CreateGrant",
+      "kms:ListGrants",
+      "kms:RevokeGrant",
+    ]
+    resources = ["*"]
+
+    principals {
+      type = "AWS"
+      identifiers = [
+        local.cluster_iam_role_pathed_arn
+      ]
+    }
+
+    condition {
+      test     = "Bool"
+      variable = "kms:GrantIsForAWSResource"
+      values   = ["true"]
+    }
+  }
+}
+
+data "aws_iam_policy_document" "assume_role_policy" {
+  statement {
+    sid     = "EKSClusterAssumeRole"
+    actions = ["sts:AssumeRole"]
+
+    principals {
+      type        = "Service"
+      identifiers = ["eks.${var.aws_context.partition_dns_suffix}"]
+    }
+  }
+}
+
+data "tls_certificate" "cluster_certificate" {
+  url = aws_eks_cluster.eks.identity[0].oidc[0].issuer
+}