+++ modules/eks/versions.tf
@@ -1,0 +1,22 @@
+terraform {
+  required_version = ">= 1.3.0"
+
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = ">= 5.60.0"
+    }
+    kubernetes = {
+      source  = "hashicorp/kubernetes"
+      version = ">= 2.10"
+    }
+    tls = {
+      source  = "hashicorp/tls"
+      version = "= 4.0.5"
+    }
+    http = {
+      source  = "hashicorp/http"
+      version = "= 3.4.3"
+    }
+  }
+}