+++ modules/k8s_eks_addons/gpu-operator.tf
@@ -13,9 +13,9 @@ resource "helm_release" "gpu_operator" {
     var.gpu_operator_config.chart_values
   ]
   timeout = 1200
+  wait    = false
   set {
     name  = "driver.version"
     value = var.gpu_operator_config.driver_version
   }
-
 }